+++ src/ui/game/components/Tabs.tsx
@@ -4,21 +4,19 @@ import { gameStore } from '../../store.js';
 import type { Argument } from '../../../action/action.js';
 import type { Box, LayoutAttributes } from '../../../board/element.js';
 
-const Drawer = ({ layout, absolutePosition, children, attributes }: {
+const Tabs = ({ layout, absolutePosition, children, attributes }: {
   layout: Partial<LayoutAttributes>,
   absolutePosition: Box,
   children: Record<string, JSX.Element[]>,
   attributes: {
-    tab?: React.ReactNode
-    closedTab?: React.ReactNode,
-    openDirection: 'up' | 'down' | 'left' | 'right',
-    openIf?: (actions: { name: string, args: Record<string, Argument> }[]) => boolean,
-    closeIf?: (actions: { name: string, args: Record<string, Argument> }[]) => boolean,
+    tabs: Record<string, React.ReactNode>
+    tabDirection: 'up' | 'down' | 'left' | 'right',
+    setTabTo?: (actions: { name: string, args: Record<string, Argument> }[]) => string,
   }
 }) => {
-  const [open, setOpen] = useState(false);
+  const [openTab, setOpenTab] = useState(Object.keys(children)[0]);
   const [pendingMoves] = gameStore(s => [s.pendingMoves]);
-  const { tab, closedTab, openIf, closeIf, openDirection } = attributes;
+  const { tabs, tabDirection, setTabTo } = attributes;
 
   const area = useMemo(() => layout?.area ?? {top: 0, left: 0, width: 100, height: 100}, [layout])
   const aspectRatio = useMemo(() => absolutePosition ? absolutePosition.width / absolutePosition.height : 1, [absolutePosition])
@@ -26,9 +24,9 @@ const Drawer = ({ layout, absolutePosition, children, attributes }: {
   useEffect(() => {
     const actions = pendingMoves?.map(m => ({ name: m.name, args: m.args })) ?? [];
 
-    if (openIf?.(actions)) setOpen(true);
-    if (closeIf?.(actions)) setOpen(false);
-  }, [openIf, closeIf, pendingMoves]);
+    const newTab = setTabTo?.(actions);
+    if (newTab) setOpenTab(newTab);
+  }, [setTabTo, setOpenTab, pendingMoves]);
 
   const style = useMemo(() => {
     return {
@@ -38,13 +36,6 @@ const Drawer = ({ layout, absolutePosition, children, attributes }: {
       width: area.width + '%',
     }
   }, [area]);
-
-  const sliderStyle = useMemo(() => {
-    return {
-      transform: `scaleX(${open || ['up', 'down'].includes(openDirection) ? 1 : 0}) scaleY(${open || ['left', 'right'].includes(openDirection) ? 1 : 0})`,
-      transformOrigin: openDirection === 'down' ? 'top' : (openDirection === 'up' ? 'bottom' : openDirection),
-    }
-  }, [openDirection, open]);
 
   /** inverse size to provide a relative box that matches the parent that the content was calculated against */
   const containerStyle = useMemo(() => {
@@ -57,56 +48,58 @@ const Drawer = ({ layout, absolutePosition, children, attributes }: {
   }, [area]);
 
   const tabStyle = useMemo(() => {
-    if (openDirection === 'down') {
+    if (tabDirection === 'down') {
       return {
-        top: `${open ? 100 : 0}%`,
+        top: `100%`,
         left: 0,
         width: `100%`,
       }
     }
-    if (openDirection === 'up') {
+    if (tabDirection === 'up') {
       return {
-        bottom: `${open ? 100 : 0}%`,
+        bottom: `100%`,
         left: 0,
         width: `100%`,
       }
     }
-    if (openDirection === 'right') {
+    if (tabDirection === 'right') {
       return {
-        left: `${open ? 100 : 0}%`,
+        left: `100%`,
         bottom: `100%`,
         width: `${100 / area.width * area.height / aspectRatio}%`,
         transform: `rotate(90deg)`,
         transformOrigin: 'bottom left',
       }
     }
-    if (openDirection === 'left') {
+    if (tabDirection === 'left') {
       return {
-        right: `${open ? 100 : 0}%`,
+        right: `100%`,
         bottom: `100%`,
         width: `${100 / area.width * area.height / aspectRatio}%`,
         transform: `rotate(-90deg)`,
         transformOrigin: 'bottom right',
       }
     }
-  }, [openDirection, aspectRatio, area, open]);
+  }, [tabDirection, aspectRatio, area]);
 
   return (
-    <div className={`drawer open-direction-${openDirection} ${open ? 'open' : 'closed'}`} style={style}>
+    <div className={`drawer open-direction-${tabDirection}`} style={style}>
       <div
-        className="drawer-tab"
+        className="drawer-tab multi"
         style={tabStyle}
-        onClick={() => setOpen(o => !o)}
       >
-        {open ? tab : closedTab ?? tab}
+        {Object.entries(tabs).map(([key, tab]) => (
+          <div key={key} className={`tabs-tab ${openTab === key ? 'active' : ''}`} onClick={() => {console.log(key);setOpenTab(key)}}>{tab}</div>
+        ))}
       </div>
-      <div className="drawer-content" style={sliderStyle}>
+      <div className="drawer-content">
         <div className="drawer-container" style={containerStyle}>
-          {children.main}
+          {openTab}
+          {children[openTab]}
         </div>
       </div>
     </div>
   );
 }
 
-export default Drawer;
+export default Tabs;